+++ src/Process.h
@@ -8,15 +8,15 @@ class Process
 {
 private:
     std::string name;
-    uint arrivalTime;
-    uint burstTime;
+    double arrivalTime;
+    double burstTime;
     uint priority;
 public:
     typedef std::shared_ptr<Process> ProcessPtr;
-    Process(std::string name, uint arrivalTime = 0, uint burstTime = 0, uint priority = 0);
+    Process(std::string name, double arrivalTime = 0, double burstTime = 0, uint priority = 0);
     std::string getName() const;
-    uint getArrivalTime() const;
-    uint getBurstTime() const;
+    double getArrivalTime() const;
+    double getBurstTime() const;
     uint getPriority() const;
 };
